requirements.txt
hydrogram==0.1.2
TgCrypto==1.2.5
psutil==5.9.8
python-dotenv==1.0.1
uvloop==0.19.0
aiofiles==23.2.1
